requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
asttokens==3.0.0
attrs==25.1.0
beautifulsoup4==4.13.3
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
decorator==5.2.1
distro==1.9.0
Django==5.1.6
djangorestframework==3.15.2
docstring_parser==0.16
duckduckgo_search==7.4.5
executing==2.2.0
frozenlist==1.5.0
genai==2.1.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.68.0
googlesearch-python==1.3.0
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
ipython==8.32.0
jedi==0.19.2
jiter==0.8.2
lxml==5.3.1
markdown-it-py==3.0.0
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.1.0
openai==0.27.10
parso==0.8.4
phi==0.6.7
phidata==2.7.10
pillow==11.1.0
prompt_toolkit==3.0.50
propcache==0.3.0
proto-plus==1.26.0
protobuf==5.29.3
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycountry==24.6.1
pydantic==2.10.6
pydantic-settings==2.8.0
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rsa==4.9
shellingham==1.5.4
smmap==5.0.2
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.6
sqlparse==0.5.3
stack-data==0.6.3
tabulate==0.9.0
tiktoken==0.3.3
tomli==2.2.1
tqdm==4.67.1
traitlets==5.14.3
typer==0.15.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
wcwidth==0.2.13
yarl==1.18.3
